requirements.txt
cachetools==3.0.0
certifi==2018.11.29
chardet==3.0.4
Click==7.0
Deprecated==1.2.4
Flask==1.0.2
google-api-python-client==1.7.7
google-auth==1.6.2
google-auth-httplib2==0.0.3
httplib2==0.12.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.0
oauth2client==4.1.3
oauthlib==3.0.0
py-trello==0.14.0
pyasn1==0.4.5
pyasn1-modules==0.2.3
PyGithub==1.43.4
PyJWT==1.7.1
python-dateutil==2.7.5
python-dotenv==0.10.1
pytz==2018.9
requests==2.21.0
requests-oauthlib==1.0.0
rsa==4.0
six==1.12.0
slackclient==1.3.0
uritemplate==3.0.0
urllib3==1.24.1
websocket-client==0.54.0
Werkzeug==0.14.1
wrapt==1.10.11
